requirements.txt
telethon
python-dotenv
pymongo
dnspython